split panodsiplay into two files for protons/gammas
